Add files via upload
PROJECT TITLE: HOTEL MANAGEMENT SYSTEM
AIM:
The project aims to convert the conventional hotel management system into a computerized system. Rather than the traditional methods that are used by the hotel for the administration, such as ledgers and bulk files, this Java project will make the hotel management system more seamless, accessible and also efficient.
TECHNOLOGIES USED:
The technologies that we used for this project are Swing, which is used to create a window-based application. MySQL is used for the database management, which is used to store the login data and verify the login data entered by the user and will allow the user to login and use the application. Swing and MySQL are used together with JAVA to develop the Hotel Management windows application.
ASSUMPTION TAKEN:
The basic assumption that made is that the user is using Windows 7 or higher. Otherwise using an open- source operating system.
Then the second assumption is that the application can only be accessed by the administrator. Only the user who is logged in through the admin details can use the options and also edit various information.
DETAILS HANDLED:
The application does the following functions:
1. Make customer form
2. Search rooms
3. Manage Staff (Add, Update, Delete and View)
4. Manage Guest (Add, Update and View)
5. Manage Room Details (Add, Update, Delete and View)
6. Manage Pricing Details (Set prices, Update and View)
SAMPLE SCENARIOS:
If this Hotel Management application is implemented for the Hotel LearnNXT. Major operations of the hotel will be automated. The system will track the room status, prices and check-in availability. If the guest needs to extend his/her stay in the hotel, the process of updating the room status and the pricing can be done with ease with the help of the application. Since the application is used by the administrator, it can be used to add, update, maintain and manage the staff details. If a single detail about a guest is updated then every thread that is linked to that particular guest will have a relative update of the information. For example, if the guest wants to extend their stay, then this change will
have its reflection in the billing section and room status. The end-user, that is a hotel administrator can
maintain, manage and monitor all of this information with ease with the help of this JAVA project of an
automated hotel management system.
REAL LIFE APPLICATIONS:
This project is based on a real-life application. This JAVA projects features high-performance speed along
with accuracy and also keeps the track of the data for the user. This project if modified with added
graphics and made more user friendly by connecting the application to a centralized server via the
internet will change the scope of the application. Because the data can be accessed, managed and
monitored by the owner from anywhere. Even though JAVA is platform-independent, to develop an
android and iOS application for this Hotel Management System made lead to some constraints. If these
platform constraints are resolved, the scope of this application would be seamless because it will allow
even the small-scale hotels to convert from the conventional methods. By changing, updating and
adjusting few classes in the program will allow the application to be used for various other managing
systems, such as Student management system, University management system, Library management
system, etc.